data_collection/requirements.txt
pyowm
pprint
simplejson
requests
softlayer-object-storage